requirements.txt
# Core Flask dependencies
flask==2.3.3
flask-cors==4.0.0

# AI/ML dependencies
openai-whisper==20231117
google-generativeai>=0.3.0

# Audio processing
torch>=2.0.0
numpy>=1.21.0
soundfile>=0.12.0
librosa>=0.10.0

# Utilities
python-dotenv>=0.19.0

# Testing dependencies (optional)
requests>=2.25.0
gtts>=2.3.0 
yagmail
pytest>=7.0.0
pytest-mock>=3.10.0
responses>=0.22.0